Weight Initialization Functions
Created weight initialization functions. This initialization is for the ReLU activation function. Other types will be added later. Actual construction of the weights will be implemented in the next commit.
Made the construction of the class simpler by only requesting the number  of nodes in each layer and calculating depth, reducing the need for manual depth input by the user through a struct.